requirements.txt
amqp==2.5.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Django==2.2.6
django-nameko==0.6.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
dnspython==1.16.0
eventlet==0.25.1
flasgger==0.9.3
Flask==1.1.1
greenlet==0.4.15
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
kombu==4.6.4
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==7.2.0
nameko==2.12.0
path.py==12.0.1
python-dotenv==0.10.3
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.5
vine==1.3.0
Werkzeug==0.16.0
wrapt==1.11.2
yagmail==0.11.220
zipp==0.6.0
